Add files via upload
I think there are a few typos in the main notebook, but I do not fully trust myself, so here is a new version in addition to an old one. Also a simple comparison of the methods added. If you find it suitable, I will modify it to have results not for one particular image but the whole dataset